requirements.txt
django~=5.2
djangorestframework~=3.16.0
djangorestframework-simplejwt~=5.5.0
django-cors-headers~=4.7.0
drf-spectacular~=0.28.0
django-polymorphic @ git+https://github.com/jazzband/django-polymorphic.git@v4.0.0a#egg=django-polymorphic
django-reversion~=5.1.0
django-admin-list-filter-dropdown~=1.0.3
drf-nested-routers~=0.94.1
django-filter~=25.1
djangorestframework-dataclasses~=1.3.1
basyx-python-sdk~=1.2.0
drf-writable-nested~=0.7.2
openai~=1.78.1
django-axes~=8.0.0
django-countries~=7.6.1
lxml<5  # Intentionally old because of basyx-python-sdk
drf_standardized_errors~=0.14.1
drf-standardized-errors[openapi]~=0.14.1